+++ src/Microsoft.TestPlatform.ObjectModel/Client/Events/DiscoveredSource.cs
@@ -1,0 +1,24 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.Serialization;
+
+#nullable disable
+
+namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Client;
+
+/// <summary>
+/// Describes a source file that was discovered, or that was requested to be discovered.
+/// </summary>
+public class DiscoveredSource
+{
+    [DataMember]
+    public string Source { get; set; }
+
+    public static List<DiscoveredSource> ToDiscoveredSources(List<string> sources)
+    {
+        return sources.Select(source => new DiscoveredSource { Source = source }).ToList();
+    }
+}
