+++ test/TranslationLayer.UnitTests/VsTestConsoleRequestSenderTests.cs
@@ -434,7 +434,7 @@ public class VsTestConsoleRequestSenderTests
 
         var mockHandler = new Mock<ITestDiscoveryEventsHandler2>();
 
-        List<string> sources = new() {   "1.dll" };
+        List<string> sources = new() { "1.dll" };
 
         var testCase = new TestCase("hello", new Uri("world://how"), source: sources[0]);
         var testsFound = new Message()
